+++ OpenSnek/Sources/OpenSnekCore/Models.swift
@@ -35,6 +35,8 @@ public enum DeviceProfileID: String, Codable, Hashable, Sendable {
     case basiliskV3Pro = "basilisk_v3_pro"
     case basiliskV335K = "basilisk_v3_35k"
     case orochiV2 = "orochi_v2"
+    case basilisk = "basilisk"
+    case lanceheadTournamentEdition = "lancehead_tournament_edition"
 }
 
 /// Defines device form factor values.